refactor(List): change __Len__ return type from uint to int
- Update __Len__() return type from uint to int in all collections
- Update ISized interface to return int
- Update Len() builtin to return int
- Update Index.Normalize, IndexExtensions, RangeExtensions to use int
- Update Slice.Normalize and Slice.Len to use int
- Update all internal usages to avoid unnecessary casts
- Update test assertions to use int literals

This change affects List<T>, Set<T>, Dict<K,V>, DictKeyView, DictItemsView,
DictValuesView, Slice, Str, and Deque.

+++ src/Sharpy.Core/Collections/Exports.cs
@@ -100,7 +100,7 @@ public class Deque<T> : IIterable<T>, ISized
         }
     }
 
-    public uint __Len__() => (uint)_list.Count;
+    public int __Len__() => _list.Count;
 
     public Iterator<T> __Iter__()
     {
