+++ src/Sharpy.Core/Partial.List/List.ISized.cs
@@ -5,8 +5,8 @@ public sealed partial class List<T>
     /// <summary>
     /// Returns the number of items in the list.
     /// </summary>
-    public uint __Len__()
+    public int __Len__()
     {
-        return (uint)_list.Count;
+        return _list.Count;
     }
 }
